Explanation of whole program
The play() function gets input from the user and from the computer by using random.choice and compares the two by using if statements. If the user and computer got the right letters the if statement will return a 'tie' expression, if there is no tie the computer goes to the next line and runs the is_win function if the player typed in the winning letter. The is_win function checks to see if the player won by going through the rules set up in the if statement and if the statement is true the function will return a True statemnt and the play() function will express 'you won' statement but if the is_win function does not return a True the play()function will return and express 'You lost' statement